ai_service/requirements.txt
# =====================================================================
# ai-service/requirements.txt — AI Microservice Dependencies
# =====================================================================

# API
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<0.31

# Data science
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scipy>=1.11,<2.0
statsmodels>=0.14,<0.15

# Validation (FastAPI uses Pydantic v2)
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0

# Networking & caching
httpx>=0.27,<0.28
redis>=5.0,<6.0

# Environment
python-dotenv>=1.0,<2.0

# Typing
typing_extensions>=4.12,<5.0

# Testing (dev / CI)
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
fakeredis>=2.23,<3.0
